Rename the 'reset_file_loaded' fixture to be descriptive and use it in tests
